Implement new Notifications Permission
Since Android 13 there is a Permission that has to be granted in order to send notifications.
I implemented it so that when the Permission is not granted, the App won't let the user add Alarms and instead show a Toast.
Also, when the App is started without notification permissions and the user has Alarms set, a Toast is displayed.